+++ internal/sdk/models/shared/assignrole.go
@@ -182,6 +182,7 @@ const (
 	EntityRegionUs       EntityRegion = "us"
 	EntityRegionEu       EntityRegion = "eu"
 	EntityRegionAu       EntityRegion = "au"
+	EntityRegionMe       EntityRegion = "me"
 	EntityRegionWildcard EntityRegion = "*"
 )
 
@@ -199,6 +200,8 @@ func (e *EntityRegion) UnmarshalJSON(data []byte) error {
 	case "eu":
 		fallthrough
 	case "au":
+		fallthrough
+	case "me":
 		fallthrough
 	case "*":
 		*e = EntityRegion(v)