+++ internal/sdk/models/shared/createapiproductdto.go
@@ -13,6 +13,13 @@ type CreateAPIProductDTO struct {
 	// Keys must be of length 1-63 characters, and cannot start with "kong", "konnect", "mesh", "kic", or "_".
 	//
 	Labels map[string]string `json:"labels,omitempty"`
+	// Public labels store information about an entity that can be used for filtering a list of objects.
+	//
+	// Public labels are intended to store **PUBLIC** metadata.
+	//
+	// Keys must be of length 1-63 characters, and cannot start with "kong", "konnect", "mesh", "kic", or "_".
+	//
+	PublicLabels map[string]string `json:"public_labels,omitempty"`
 	// The list of portal identifiers which this API product should be published to
 	PortalIds []string `json:"portal_ids"`
 }
@@ -36,6 +43,13 @@ func (o *CreateAPIProductDTO) GetLabels() map[string]string {
 		return nil
 	}
 	return o.Labels
+}
+
+func (o *CreateAPIProductDTO) GetPublicLabels() map[string]string {
+	if o == nil {
+		return nil
+	}
+	return o.PublicLabels
 }
 
 func (o *CreateAPIProductDTO) GetPortalIds() []string {