+++ internal/sdk/models/shared/vault.go
@@ -5,7 +5,7 @@ package shared
 // Vault entities are used to configure different Vault connectors. Examples of Vaults are Environment Variables, Hashicorp Vault and AWS Secrets Manager. Configuring a Vault allows referencing the secrets with other entities. For example a certificate entity can store a reference to a certificate and key, stored in a vault, instead of storing the certificate and key within the entity. This allows a proper separation of secrets and configuration and prevents secret sprawl.
 type Vault struct {
 	// The configuration properties for the Vault which can be found on the vaults' documentation page.
-	Config any `json:"config,omitempty"`
+	Config any `json:"config"`
 	// Unix epoch when the resource was created.
 	CreatedAt *int64 `json:"created_at,omitempty"`
 	// The description of the Vault entity.