+++ tests/e2e/api-product.tf
@@ -3,7 +3,7 @@ resource "konnect_portal" "my_portal" {
   name                      = "My New Portal"
   auto_approve_applications = false
   auto_approve_developers   = false
-  custom_domain             = "tfdemo.example.com"
+  custom_domain             = "tf-example-domain.example.com"
   is_public                 = false
   rbac_enabled              = false
 }