+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 2.1.0
+> Released on 2024/12/06
+
+### Features
+
+* Add support for `konnect_gateway_mtls_auth` resource
+* API Products now support `public_labels`
+* Added support for Konnect Middle-East region
+
+### Bug Fixes
+
+* Fix `Expected #sdk.Konnect, got: *sdk.Konnect.` error with custom plugins (and add an integration test)
+
 ## 2.0.2
 > Released on 2024/11/12